Contact form (#37)
* Run migrations after successful deployment

* Index berichten

The index page for the contacts messages. First attemp to show the active item in the sidebar.

* Sidenav + Details page

Changed the sidenav so the it show the curretn active page. Changed _context to _repository.

* Added new prop to Contact Model

* Badge shows current unread messages

change badge css to display none

* Changed breadcrumb

* Test not working

* Changed Repository & Unit test

Not working yet.

* Made 6 Unit test for the detaisl action

They all pass :+1:  plus the index unit test. Changed ajax to make it more 'perfect'

* Fixed the test

As requested by michael. I made a constructor wich create a mock so I can reuse that one mock! :+1:

* Little change